Major overhaul to resource loading and structure
Code structure is separated in src/ and include/ directories now for source and header files respectively.
All resources are now loaded in similar ways, with the maps and textures being compressed, which lowers the executable size in general.
Also, now using Tiled to generate the game maps, separating the map in Walls, Floor and Ceiling layers.
A new helper executable was created to aid in compiling everything correctly with a auto-generated makefile.